requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.14
Django==4.1
django-elasticsearch-dsl==7.2.2
django-filter==22.1
djangorestframework==3.13.1
elasticsearch==7.14.0
elasticsearch-dsl==7.4.0
kafka-python==2.0.2
Pillow==9.2.0
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
